requirements.txt
pandas==3.0.1
prefect==3.4.19
psycopg2-binary==2.9.10
pytest==8.4.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
